Add: new interfaces for beamInput and beamOutput
