+++ src/app/beamInterfaces/ibeaminput.ts
@@ -1,0 +1,5 @@
+export interface IBeamInput {
+    l: number;
+    a: number;
+    w: number;
+  }
