+++ src/app/beamInterfaces/ibeamoutput.ts
@@ -1,0 +1,9 @@
+export interface IBeamOutput {
+    va: number;
+    vb1: number;
+    vb2: number;
+    r1: number;
+    r2: number;
+    msagMax: number;
+    mhogMax: number;
+  };
