+++ src/app/beamInterfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './ibeaminput';
+export * from './ibeamoutput';
